requirements.txt
python-telegram-bot==22.3
python-telegram-bot[rate-limiter]
pycoingecko==3.1.0
requests>=2.31
numpy>=1.26
pandas>=2.2
matplotlib>=3.8
python-dotenv>=1.0
tzlocal==2.1
apscheduler==3.10.4
pytz==2024.1
pillow
